READY FOR REVIEW (fundamental good exercise)
tbh i kept too many iterations of the code that i couldnt find which was my latest (and most well done) code. took some time to find it but then i added some comments to the incorrect codes to note what i learnt. noted down some questions in a markdown cell too! honestly a bit confused cos python seems to accept some incorrect codes sometimes but not other times.. but YESSS it was so thrilling to finally improve the code to be able to recognise negative numbers with non-zero decimals and type out an error msg specifying that :)